Moving URL filters and URL sanitizers to their own namespaces
