+++ Lombiq.Tests.UI/MonkeyTesting/MonkeyTestingOptions.cs
@@ -1,3 +1,5 @@
+using Lombiq.Tests.UI.MonkeyTesting.UrlFilters;
+using Lombiq.Tests.UI.MonkeyTesting.UrlSanitizers;
 using System;
 using System.Collections.Generic;
 using System.Security.Cryptography;
