+++ Lombiq.Tests.UI/MonkeyTesting/UrlFilters/MatchesRegexMonkeyTestingUrlFilter.cs
@@ -2,7 +2,7 @@ using Lombiq.Tests.UI.Services;
 using System;
 using System.Text.RegularExpressions;
 
-namespace Lombiq.Tests.UI.MonkeyTesting
+namespace Lombiq.Tests.UI.MonkeyTesting.UrlFilters
 {
     /// <summary>
     /// URL filter that matches the URL against the configured regex pattern.
