+++ Lombiq.Tests.UI/MonkeyTesting/UrlFilters/StartsWithMonkeyTestingUrlFilter.cs
@@ -1,6 +1,6 @@
 using Lombiq.Tests.UI.Services;
 using System;
-namespace Lombiq.Tests.UI.MonkeyTesting
+namespace Lombiq.Tests.UI.MonkeyTesting.UrlFilters
 {
     /// <summary>
     /// URL filter that matches the URL to see if starts with the configured relative URL.
