+++ Lombiq.Tests.UI/MonkeyTesting/UrlSanitizers/RemovesBaseUrlMonkeyTestingUrlSanitizer.cs
@@ -1,7 +1,7 @@
 using Lombiq.Tests.UI.Services;
 using System;
 
-namespace Lombiq.Tests.UI.MonkeyTesting
+namespace Lombiq.Tests.UI.MonkeyTesting.UrlSanitizers
 {
     /// <summary>
     /// Represents the URL sanitizer that removes an <see cref="UITestContext"/> base URL part if it is present.
