+++ Lombiq.Tests.UI/MonkeyTesting/UrlSanitizers/RemovesByRegexMonkeyTestingUrlSanitizer.cs
@@ -2,7 +2,7 @@ using Lombiq.Tests.UI.Services;
 using System;
 using System.Text.RegularExpressions;
 
-namespace Lombiq.Tests.UI.MonkeyTesting
+namespace Lombiq.Tests.UI.MonkeyTesting.UrlSanitizers
 {
     /// <summary>
     /// URL sanitizer that removes parts that match the specific regex pattern.
