+++ Lombiq.Tests.UI/MonkeyTesting/UrlSanitizers/RemovesQueryParameterMonkeyTestingUrlSanitizer.cs
@@ -1,4 +1,4 @@
-namespace Lombiq.Tests.UI.MonkeyTesting
+namespace Lombiq.Tests.UI.MonkeyTesting.UrlSanitizers
 {
     /// <summary>
     /// URL sanitizer that removes specific query parameter.
